move view to MovieDatails
